requirements.txt
streamlit
requests
pymupdf
pdfplumber
faiss-cpu
numpy
requests
rank-bm25
sentence-transformers
torch
langchain
langchain-groq

